remove code duplication in getUserSyncs func for oms, admatic, rubicon and dianomi adapters
