+++ modules/dianomiBidAdapter.js
@@ -15,6 +15,7 @@ import {
 import { config } from '../src/config.js';
 import { Renderer } from '../src/Renderer.js';
 import { convertOrtbRequestToProprietaryNative } from '../src/native.js';
+import {getUserSyncParams} from '../libraries/userSyncUtils/userSyncUtils.js';
 
 const { getConfig } = config;
 
@@ -302,19 +303,8 @@ export const spec = {
       .filter(Boolean);
   },
   getUserSyncs: (syncOptions, responses, gdprConsent, uspConsent) => {
-    const params = {};
-    if (gdprConsent) {
-      if (typeof gdprConsent.gdprApplies === 'boolean') {
-        params['gdpr'] = Number(gdprConsent.gdprApplies);
-      }
-      if (typeof gdprConsent.consentString === 'string') {
-        params['gdpr_consent'] = gdprConsent.consentString;
-      }
-    }
+    const params = getUserSyncParams(gdprConsent, uspConsent);
 
-    if (uspConsent) {
-      params['us_privacy'] = encodeURIComponent(uspConsent);
-    }
     if (syncOptions.iframeEnabled) {
       // data is only assigned if params are available to pass to syncEndpoint
       return {
